+++ tests/table/apis/table.entity.tests.azure.data-tables.ts
@@ -6,10 +6,9 @@ import {
   odata,
   TableEntity,
   TableClient,
-  AzureNamedKeyCredential,
-  TableTransaction,
-  TransactionAction
+  TableTransaction
 } from "@azure/data-tables";
+import { AzureNamedKeyCredential } from "@azure/core-auth";
 import { configLogger } from "../../../src/common/Logger";
 import TableServer from "../../../src/table/TableServer";
 import {
@@ -35,10 +34,15 @@ describe("table Entity APIs test", () => {
   let server: TableServer;
   const tableName: string = getUniqueName("datatables");
 
+  const sharedKeyCredential = new AzureNamedKeyCredential(
+    EMULATOR_ACCOUNT_NAME,
+    EMULATOR_ACCOUNT_KEY
+  );
+
   const tableClient = new TableClient(
     `https://${HOST}:${PORT}/${EMULATOR_ACCOUNT_NAME}`,
     tableName,
-    new AzureNamedKeyCredential(EMULATOR_ACCOUNT_NAME, EMULATOR_ACCOUNT_KEY)
+    sharedKeyCredential
   );
 
   const requestOverride = { headers: {} };
@@ -59,18 +63,18 @@ describe("table Entity APIs test", () => {
   it("Batch API should return row keys in format understood by @azure/data-tables, @loki", async () => {
     await tableClient.createTable();
     const partitionKey = createUniquePartitionKey();
-    const testEntities: any[] = [
+    const testEntities: AzureDataTablesTestEntity[] = [
       createBasicEntityForTest(partitionKey),
       createBasicEntityForTest(partitionKey),
       createBasicEntityForTest(partitionKey)
     ];
-
-    const batch: TransactionAction[] = [];
-    for (const entity of testEntities) {
-      batch.push(["create", entity]);
+    const transaction = new TableTransaction();
+    for (const testEntity of testEntities) {
+      transaction.createEntity(testEntity);
     }
 
-    const result = await tableClient.submitTransaction(batch);
+    const result = await tableClient.submitTransaction(transaction.actions);
+
     assert.ok(result.subResponses[0].rowKey);
     await tableClient.deleteTable();
   });
@@ -79,9 +83,8 @@ describe("table Entity APIs test", () => {
   it("Batch API should correctly process LogicApp style update request sequence", async () => {
     await tableClient.createTable();
     const logicAppReproEntity = new LogicAppReproEntity();
-    const insertedEntityHeaders = await tableClient.createEntity<LogicAppReproEntity>(
-      logicAppReproEntity
-    );
+    const insertedEntityHeaders =
+      await tableClient.createEntity<LogicAppReproEntity>(logicAppReproEntity);
     assert.notStrictEqual(insertedEntityHeaders.etag, undefined);
     logicAppReproEntity.sequenceNumber = 1;
     logicAppReproEntity.testString = "1";
@@ -112,12 +115,14 @@ describe("table Entity APIs test", () => {
     updatedEntity.sequenceNumber = 2;
 
     // Batch using replace mode
-    const batch1 = new TableTransaction();
-    batch1.createEntity(batchEntity1);
-    batch1.createEntity(batchEntity2);
-    batch1.updateEntity(updatedEntity, "Replace");
+    const transaction = new TableTransaction();
 
-    const result = await tableClient.submitTransaction(batch1.actions);
+    transaction.createEntity(batchEntity1);
+    transaction.createEntity(batchEntity2);
+    transaction.updateEntity(updatedEntity, "Replace");
+
+    const result = await tableClient.submitTransaction(transaction.actions);
+
     // batch operations succeeded
     assert.strictEqual(
       result.subResponses[0].status,
@@ -137,11 +142,12 @@ describe("table Entity APIs test", () => {
     // we have a new etag from the updated entity
     assert.notStrictEqual(result.subResponses[2].etag, updatedEntity.etag);
 
-    const batch2 = new TableTransaction();
-    batch2.deleteEntity(batchEntity1.partitionKey, batchEntity1.rowKey);
-    batch2.deleteEntity(batchEntity2.partitionKey, batchEntity2.rowKey);
+    const transaction2 = new TableTransaction();
+    transaction2.deleteEntity(batchEntity1.partitionKey, batchEntity1.rowKey);
+    transaction2.deleteEntity(batchEntity2.partitionKey, batchEntity2.rowKey);
+    transaction2.deleteEntity(updatedEntity.partitionKey, updatedEntity.rowKey);
 
-    const result2 = await tableClient.submitTransaction(batch2.actions);
+    const result2 = await tableClient.submitTransaction(transaction2.actions);
     assert.strictEqual(
       result2.subResponses[0].status,
       204,
@@ -164,9 +170,8 @@ describe("table Entity APIs test", () => {
   it("Should return bad request error for incorrectly formatted etags, @loki", async () => {
     await tableClient.createTable();
     const partitionKey = createUniquePartitionKey();
-    const testEntity: AzureDataTablesTestEntity = createBasicEntityForTest(
-      partitionKey
-    );
+    const testEntity: AzureDataTablesTestEntity =
+      createBasicEntityForTest(partitionKey);
 
     const result = await tableClient.createEntity(testEntity);
 
@@ -209,9 +214,8 @@ describe("table Entity APIs test", () => {
 
   it("should find an int as a number, @loki", async () => {
     const partitionKey = createUniquePartitionKey();
-    const testEntity: AzureDataTablesTestEntity = createBasicEntityForTest(
-      partitionKey
-    );
+    const testEntity: AzureDataTablesTestEntity =
+      createBasicEntityForTest(partitionKey);
 
     await tableClient.createTable({ requestOptions: { timeout: 60000 } });
     const result = await tableClient.createEntity(testEntity);
@@ -230,9 +234,8 @@ describe("table Entity APIs test", () => {
 
   it("should find a long int, @loki", async () => {
     const partitionKey = createUniquePartitionKey();
-    const testEntity: AzureDataTablesTestEntity = createBasicEntityForTest(
-      partitionKey
-    );
+    const testEntity: AzureDataTablesTestEntity =
+      createBasicEntityForTest(partitionKey);
 
     await tableClient.createTable({ requestOptions: { timeout: 60000 } });
     const result = await tableClient.createEntity(testEntity);
@@ -252,9 +255,8 @@ describe("table Entity APIs test", () => {
 
   it("should find an entity using a partition key with multiple spaces, @loki", async () => {
     const partitionKey = createUniquePartitionKey() + " with spaces";
-    const testEntity: AzureDataTablesTestEntity = createBasicEntityForTest(
-      partitionKey
-    );
+    const testEntity: AzureDataTablesTestEntity =
+      createBasicEntityForTest(partitionKey);
 
     await tableClient.createTable({ requestOptions: { timeout: 60000 } });
     const result = await tableClient.createEntity(testEntity);